requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
django-filter>=24.0
django-storages>=1.14
boto3>=1.34
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
mutagen>=1.47
pillow>=10.0
gunicorn>=22.0
psycopg2-binary>=2.9
whitenoise>=6.6
dj-database-url>=2.1
librosa>=0.10
